Avail setup.
- Added `avail`, `avail-stdlib`, and ``avail-plugin` to `libs.versions.toml`.
- Updated `lang/build.gradle.kts` to:
  - Add Avail support.
  - Remove unused dependencies.
- Added Gradle task `:lang:launchAvail` to launch Anvil (the Avail IDE) on
  the Avail project rooted at `lang`.
- Likewise, added an IntelliJ run configuration to launch Anvil (for
  convenience).
- Added an Avail root `xef` to host the prototype DSL.
  - Introduced `xef/Compiler` as a placeholder for the DSL compiler.

When Anvil is launched for the first time, you will need to use it to retrieve
the Avail standard library (≥ `2.0.0.alpha22-1.6.1.alpha11`). The project
manager window (`Avail Projects`) opens, and there is a `Settings` button at
the bottom. Pressing it opens a `Settings` window, and on the left is
`Standard Library`. Clicking here will cause Anvil to search Maven Central for
the latest version of the Avail standard library. Install this version.
Afterward, you can double-click `xef` in the project manager window to open
the Anvil Workbench on the `xef` project. Visiting the `Standard Library`
view on subsequent occasions provides an opportunity to download newer
versions.