Implemented constant and variable declarations.
- Supported primitive types are:
  - `Boolean` (Aristotelian bivalent logic)
  - `Int` (64-bit 2's complement integer)
  - `Float` (IEEE 754 double-precision floating point numbers)
  - `String` (Unicode text)
  - `Type` (metatype encompassing all types)
- Tuple and labeled tuple are supported in the model.
  - Labeled tuple is not yet supported syntactically.
- Constant and variable declarations are supported.
  - Tuple type declarations do not yet support intermixing non-type values and types.
  - But actual tuples may mix types and values freely.
- Updated to the latest version of the Avail toolchain.

I've put support for significant whitespace on the back burner, because we are going
to handle it via a new core parsing mechanism in Avail. @markATavail is leading the
implementation, and I am assisting with specification, implementation, testing, and
review.

If anyone wants to play with the declaration syntax, just start a new branch (for
safety) and add your declarations after the `Body` keyword in `Compiler.avail`. The
module header (ending with `Body`) provides the complete language specification for
the nascent implemented version of the Xef prompting language, and over time will
grow into the complete technical specification. It will be analogous to an EBNF, but
using a more powerful Avail-native syntax.